api/pyproject.toml
[project]
name = "openapi-to-mcp-backend"
version = "0.1.0"
description = "FastAPI backend for OpenAPI to MCP converter"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.31.1",
    "pydantic>=2.12.0",
    "pydantic-settings>=2.11.0",
    "anthropic>=0.69.0",
    "pyyaml>=6.0.3",
    "jinja2>=3.1.6",
    "python-multipart>=0.0.20",
    "python-dotenv>=1.1.1",
    "mcp>=1.17.0",
    "httpx>=0.28.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = []
